chore: Fixed missing section action preview
