+++ test/components/previews/viral_card_component_preview/section_with_action.html.erb
@@ -1,6 +1,6 @@
-<%= viral_card(title: "Card with multiple titled sections") do |card| %>
-  <% card.with_section(title: "First Sections", border_top: true) do |section| %>
-    <% section.action do %>
+<%= viral_card(title: "Card with section actions") do |card| %>
+  <% card.with_section(title: "First Section", border_top: true) do |section| %>
+    <% section.with_action do %>
       <%= link_to "Delete", "#", class: "text-red-500 dark:text-red-400" %>
     <% end %>
     <p class="text-base text-gray-900 dark:text-white">First Section Content</p>
