+++ test/components/previews/viral_card_component_preview/with_multiple_sections.html.erb
@@ -1,10 +1,8 @@
 <%= viral_card(title: "Card with multiple sections") do |card| %>
-  <% card.with_section(border_bottom: true, border_top: true) do %>
-    <h1 class="text-3xl font-bold dark:text-white">Section 1</h1>
+  <% card.with_section(border_bottom: true, border_top: true, title: "Section 1") do %>
     <p class="text-base text-gray-900 dark:text-white">Some content</p>
   <% end %>
-  <% card.with_section do %>
-    <h1 class="text-3xl font-bold dark:text-white">Section 2</h1>
+  <% card.with_section(title: "Section 2") do %>
     <p class="text-base text-gray-900 dark:text-white">Some content</p>
   <% end %>
 <% end %>
